Added Data Models in ApplicationDbContext.cs
made GeneralCurrencyId as nullable in Customer.cs
Added models for WorkOrder , WorkOrderLabor , WorkOrderPartNumber , WorkOrderStage , WorkOrderStatus , WorkOrderType
Added a new method in CustomerRepository to get the Customer specific information.
Added a new method in EmployeeRepository to get the Employee Specific information.
Added a method for Implementattion in ICustomerRepository.
Added a method for Implementation in IEmployee.
Imported the service in app.module.ts.
Added TypeScripts models for the same Database models.